Change shopping item hierarchy
Change interfaces (so loose design) to abstract class since they define
the same type
