+++ src/Demo/CartBundle/Entity/Cart/AbstractCartManager.php
@@ -11,7 +11,7 @@ namespace Demo\CartBundle\Entity\Cart;
 
 use Demo\CartBundle\Entity\OrderElement\OrderItemInterface;
 use Demo\CartBundle\Entity\OrderElement\WishOrderItemInterface;
-use Demo\CartBundle\Entity\Product\ShoppingItemInterface;
+use Demo\CartBundle\Entity\Product\AbstractShoppingItem;
 use Doctrine\Common\Persistence\ObjectManager;
 
 /**
@@ -61,7 +61,7 @@ abstract class AbstractCartManager implements CartManagerInterface
     /**
      * @inheritDoc
      */
-    public abstract function addShoppingItem(ShoppingItemInterface $item);
+    public abstract function addShoppingItem(AbstractShoppingItem $item);
 
     /**
      * @inheritDoc
