+++ src/Demo/CartBundle/Entity/Product/ShoppingItem.php
@@ -16,109 +16,8 @@ use Doctrine\ORM\Mapping as ORM;
  * @ORM\DiscriminatorColumn(name="discr", type="string")
  * @ORM\DiscriminatorMap({"normal" = "ShoppingItem", "sale" = "SaleItem"})
  */
-class ShoppingItem implements ShoppingItemInterface
+class ShoppingItem extends AbstractShoppingItem
 {
-    /**
-     * @ORM\Column(type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    protected $name;
-
-    /**
-     * @ORM\Column(type="text")
-     */
-    protected $description;
-
-    /**
-     * @ORM\Column(type="decimal", scale=2)
-     */
-    protected $price;
-
-    /**
-     * @ORM\Column(type="boolean")
-     */
-    protected $inStock;
-
-    /**
-     * @return mixed
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * @param mixed $name
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getDescription()
-    {
-        return $this->description;
-    }
-
-    /**
-     * @param mixed $description
-     */
-    public function setDescription($description)
-    {
-        $this->description = $description;
-    }
-
-    /**
-     * returns the actual price of the item
-     *
-     * @return mixed
-     */
-    public function getPrice()
-    {
-        return $this->price;
-    }
-
-    /**
-     * @param mixed $price
-     */
-    public function setPrice($price)
-    {
-        $this->price = $price;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getInStock()
-    {
-        return $this->inStock;
-    }
-
-    /**
-     * @param mixed $inStock
-     */
-    public function setInStock($inStock)
-    {
-        $this->inStock = $inStock;
-    }
-
     /**
      * returns the net price of an item after discount(if exists) or other item price manipulation
      *
